:sparkles: isLoggedIn에 따른 북마크, 좋아요 조건 처리

+++ components/button/BookmarkButton.tsx
@@ -1,31 +1,40 @@
 import React, { useCallback, useEffect } from 'react';
 import imgBookmark from '/images/icon-bookmark.svg';
-import { useSelector, useDispatch } from 'react-redux';
-import type { RootState } from '../../renderer/store.js';
+import { useDispatch } from 'react-redux';
+import { API_URL } from '../../pages/API_URL/api';
+import { navigate } from 'vite-plugin-ssr/client/router';
+import { useAppSelector } from '../../renderer/store/hooks';
 
 export { BookmarkButton };
 
-function BookmarkButton(props: { restaurantId: string }) {
+function BookmarkButton({ restaurantId, isLoggedIn }: { restaurantId: string; isLoggedIn: boolean }) {
     const dispatch = useDispatch();
-    const hasBookmarkList = useCallback(async () => {
-        const res = await fetch(`http://localhost:5000/api/users/`);
-        const data = await res.json();
 
-        return data.user.bookmarkList.some((v: { _id: string; registeredAt: string }) => props.restaurantId === v._id);
-    }, [props.restaurantId]);
+    const hasBookmarkList = useCallback(async () => {
+        if (isLoggedIn) {
+            const res = await fetch(`${API_URL}/users/`);
+            const data = await res.json();
+
+            return data.user.bookmarkList.some((v: { _id: string; registeredAt: string }) => {
+                return restaurantId === v._id;
+            });
+        }
+    }, [restaurantId, isLoggedIn]);
 
     const delBookmark = useCallback(async () => {
-        const res = await fetch(`http://localhost:5000/api/users/bookmark/${props.restaurantId}`, {
-            method: 'DELETE',
-        });
-
-        if (res.ok) {
-            dispatch({
-                type: 'buttonSlice/HAS_BOOKMARK_LIST',
-                ON: false,
+        if (isLoggedIn) {
+            const res = await fetch(`${API_URL}/users/bookmark/${restaurantId}`, {
+                method: 'DELETE',
             });
-        } else throw new Error();
-    }, [props.restaurantId]);
+
+            if (res.ok) {
+                dispatch({
+                    type: 'buttonSlice/HAS_BOOKMARK_LIST',
+                    ON: false,
+                });
+            } else console.error('북마크 취소에 실패했습니다.');
+        }
+    }, [restaurantId, isLoggedIn]);
 
     useEffect(() => {
         hasBookmarkList().then((bool) => {
@@ -34,22 +43,28 @@ function BookmarkButton(props: { restaurantId: string }) {
                 ON: bool,
             });
         });
-    }, [hasBookmarkList]);
+    }, [hasBookmarkList, isLoggedIn]);
 
     function handleClick() {
-        hasBookmarkList().then((bool) => {
-            return bool ? delBookmark() : addBookmark();
-        });
+        if (isLoggedIn) {
+            hasBookmarkList().then((bool) => {
+                return bool ? delBookmark() : addBookmark();
+            });
+        } else {
+            if (confirm('로그인이 필요한 서비스입니다.\n로그인 하시겠습니까?')) {
+                navigate('/login');
+            }
+        }
     }
 
     async function addBookmark() {
         try {
-            const res = await fetch(`http://localhost:5000/api/users/bookmark`, {
+            const res = await fetch(`${API_URL}/users/bookmark`, {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
                 },
-                body: JSON.stringify({ _id: props.restaurantId }),
+                body: JSON.stringify({ _id: restaurantId }),
             });
 
             if (res.ok) {
@@ -68,7 +83,7 @@ function BookmarkButton(props: { restaurantId: string }) {
             <img
                 src={imgBookmark}
                 alt="북마크 이미지"
-                className={useSelector((state: RootState) => {
+                className={useAppSelector((state) => {
                     return state.buttonSlice.bookmark.ON ? 'img-bookmark on' : 'img-bookmark';
                 })}
             />
